+++ main.py
@@ -28,35 +28,38 @@ def main() -> None:
     logging.info("Hello NeuraKitten... 🐾\n")
 
     cfg = NeuraConfig(
-        epochs=5000,
-        hidden_layers=[32, 32, 32, 32],
-        samples=2000,
+        epochs=1000,
+        hidden_layers=[128,96,64],
+        samples=5000,
         batch_size=50,
         balanced_batches=True,
         data_mode="spirals",  # "multidonut", "spirals", "rhodonea", "iris"
         # mdonut_r_evenly_dist=False,
         # mdonut_radii=[4, 3, 2, 1],
         spiral_max_radius=1,
-        spiral_num_classes=3,
-        spiral_turns=7,
-        noise=0.03,
+        spiral_num_classes=5,
+        spiral_turns=3,
+        noise=0.02,
         feature_mode="cartesian",
-        initial_lr=0.002,
-        decay_rate=0.001,
-        use_interaction=True,
-        use_squares=True,
-        use_trig=True,
-        visualize=True,
-        frame_log=50,
+        initial_lr=0.001,
+        decay_rate=0.000,
+        use_interaction=False,
+        use_squares=False,
+        use_trig=False,
+        visualize=False,
+        frame_log=100,
         frame_visual=1,
         save_to_file=True,
-        checkpoint_interval=50,
+        checkpoint_interval=1,
     )
 
     pipeline = NeuraPipeline(experiment_name="NeuraKitten_Standard", cfg=cfg)
 
     try:
-        pipeline.run()
+        pipeline.run(
+            mode="replay",
+            experiment_path="experiments/20260803_173932_NeuraKitten_Standard",
+        )
         logging.info("\nExperiment completed successfully.")
     except KeyboardInterrupt:
         logging.info("\nExperiment interrupted by user. Cleaning up...")